WIP: 페어코딩 준비
